Refactor project file structure
