+++ cg-app/src/assets/styles/index.css
